Fix hud hotloading by just making hud a RootPanel, not a HudEntity
Because of this, the hud doesn't need to be an entity anymore, meaning it can be instantiated fully on the client. Matt (fp employee) brought up a good point about S->C RPC stuff, but I think that can be handled with Events.

Also the background of the number displays was made a bit darker.