Take devcam initial settings from local pawn's camera, not CurrentView
If the FOV of the copied camera is 0, then the value for FovOverride will be taken from default_fov. A value of 0 for the FOV visually works, but the movement code no longer functions due to 0 being what it is.

Code was also added for the same reason to the ortho toggle, it estimates an ortho size for the current FOV.