Add IACarriable and a flashlight
Because IACarriable is the new base carriable class, it removes a lot from IAWeapon and centralizes it all there. IATool also inheirts from IACarriable but it's a stub right now.

When making the flashlight, I was trying to solve dropping not working for nearly 2 hours until I realized the pickup trigger was just constantly going off. SleepingPickupTrigger was added so that carriables cannot be picked up for a certain amount of time after they are dropped.