HUGE WEAPON AND DEBUGGING REFACTOR
Left this alone for 3 months and FP ended up fixing some of the bugs that were holding progress back, so here's a lot of previously uncommitted changes all at once.

Weapons were refactored to be part of the shared CK library, and all debugging was moved there as well.

The base types (Carriable, Tool, WeaponFirearm) now have three stages of implementation: interface, definition, and entity. Interface unsurprisingly defines the core things the types need, the definition just provides a GameResource to the game, and the entity is where all the actual implementation happens.

It's been separated like this to allow for changes outside the use of definitions, essentially the definition is now only a mold and not the definite rule for weapon properties.

Rounds are also planned to be separated like this but there's some jankiness happening wrt referencing the asset in a stuct.