Player refactor: ArcadePlayer to CKPlayer
Big refactor to the way the player class works. Most of the stuff that was implemented into ArcadePlayer and its friends (armor, clothes, the inventory, cameras, animators, QPhys) could absolutely be stripped out and thrown into the CK shared library.

So it was! All of that has been moved and tweaked a bit. It hasn't been thoroughly tested so bugs are probably bound to occur, but for now everything seems to be working fine.